refactor(utils): remove barrel files from utils pkg
