+++ packages/cli/src/commands/scaffold.ts
@@ -1,5 +1,5 @@
 import { launchScaffold } from "@bluecadet/launchpad-scaffold";
-import { LogManager } from "@bluecadet/launchpad-utils";
+import { LogManager } from "@bluecadet/launchpad-utils/log-manager";
 import type { GlobalLaunchpadArgs } from "../cli.js";
 
 export async function scaffold(_argv: GlobalLaunchpadArgs) {
