+++ packages/content/src/sources/strapi-source.ts
@@ -1,4 +1,4 @@
-import type { Logger } from "@bluecadet/launchpad-utils";
+import type { Logger } from "@bluecadet/launchpad-utils/log-manager";
 import ky from "ky";
 import qs from "qs";
 import { z } from "zod";
