+++ packages/controller/src/ipc/ipc-client.ts
@@ -4,7 +4,8 @@
  */
 
 import net from "node:net";
-import type { BaseCommand, LaunchpadEvents } from "@bluecadet/launchpad-utils";
+import type { BaseCommand } from "@bluecadet/launchpad-utils/controller-interfaces";
+import type { LaunchpadEvents } from "@bluecadet/launchpad-utils/types";
 import { applyPatches, enablePatches, type Patch } from "immer";
 import { errAsync, okAsync, ResultAsync } from "neverthrow";
 import { EventBus } from "../core/event-bus.js";
