+++ packages/monitor/src/core/__tests__/process-manager.test.ts
@@ -1,5 +1,5 @@
 import { createMockLogger } from "@bluecadet/launchpad-testing/test-utils.ts";
-import type { Logger } from "@bluecadet/launchpad-utils";
+import type { Logger } from "@bluecadet/launchpad-utils/log-manager";
 import pm2 from "pm2";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import { ProcessManager } from "../process-manager.js";
