+++ packages/monitor/src/utils/__tests__/sort-windows.test.ts
@@ -1,6 +1,6 @@
 import { afterEach } from "node:test";
 import { createMockLogger } from "@bluecadet/launchpad-testing/test-utils.ts";
-import type { Logger } from "@bluecadet/launchpad-utils";
+import type { Logger } from "@bluecadet/launchpad-utils/log-manager";
 import { type Window, windowManager } from "node-window-manager";
 import semver from "semver";
 import { beforeEach, describe, expect, it, vi } from "vitest";
