+++ packages/monitor/src/utils/sort-windows.ts
@@ -1,4 +1,4 @@
-import type { Logger } from "@bluecadet/launchpad-utils";
+import type { Logger } from "@bluecadet/launchpad-utils/log-manager";
 import chalk from "chalk";
 import { windowManager } from "node-window-manager";
 import semver from "semver";
